% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{as.las}
\alias{as.las}
\title{Coherce Data Frame To ´las´ Class}
\usage{
as.las(x, id = "WELL")
}
\arguments{
\item{x}{a data frame}

\item{id}{character string to be used as ID. If it is any of the columns names of \code{x},
it used the unique of that columen as is. A different string can be passed. Default = "WELL"}
}
\description{
The function takes a data frame and try to coherce it to ´las´ R6 Class, building the corresponding
   header and tracks information.
}
