% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{colnames.las}
\alias{colnames.las}
\title{colnames() S3 Method - R6 Method caller}
\usage{
colnames.las(x, do.NULL = TRUE, prefix = "col")
}
\arguments{
\item{x}{las object}

\item{do.NULL}{logical. If \code{FALSE} and names are \code{NULL}, names are created}

\item{prefix}{for created names.}
}
\description{
colnames() S3 Method - R6 Method caller
}
