% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{extract_lasinfo}
\alias{extract_lasinfo}
\title{Extract Info From .las Line}
\usage{
extract_lasinfo(oneline, key)
}
\arguments{
\item{oneline}{character string containing the info to be extracted from .las file}

\item{key}{character string defining the keyword of the info to be extracted}
}
\value{
string with the extracted info value
}
\description{
Extract Info From .las Line
}
\note{
internal function used in \code{read_las}
}
