% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petroreadr.R
\name{extract_wellname}
\alias{extract_wellname}
\title{Extract Well Name from .las line}
\usage{
extract_wellname(oneline)
}
\arguments{
\item{oneline}{string character containing the well name extracted form .las file}
}
\value{
string with the ecxtacte well name
}
\description{
Extract Well Name from .las line
}
\note{
internal function used in \code{read_las}
}
