% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{is.las}
\alias{is.las}
\title{Check If The Argument Belongs To ´las´ R6 Class}
\usage{
is.las(x)
}
\arguments{
\item{x}{a R object}
}
\value{
logical value, \code{TRUE} or \code{FALSE}
}
\description{
Check If The Argument Belongs To ´las´ R6 Class
}
