% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\docType{data}
\name{las}
\alias{las}
\title{`las` Class Definition (R6)}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
las
}
\description{
The class `las` provides customized behaviour of the following functions:
\itemize{
  \item{\code{print()}} {print method for `las` objects}
  \item{\code{summary()}}
  \item{\code{head()} / \code{tail()} / \code{dim()} / \code{scale()}, \code{colnames()}} {work directly on \code{obj$data} as it is a data frame}
  \item{\code{print()}} {work directly on \code{is_numeric(obj$data)}, applying the function across each column}
  \item{\code{print()}} {work directly on \code{obj$data} as it is a data frame, plus modifying accordigly \code{obj$tracks}}
}
}
\details{
Additional methods are:
\itemize{
  \item{\code{set_header()}} {set values or new lines on \code{obj$header}}
  \item{\code{set_tracks()}} {set values or new lines on \code{obj$tracks}: if a new line is added, then this track is automatically
        added to \code{obj$data} as \code{NA}}
}
}
\section{Fields}{

\describe{
\item{\code{header}}{General information from the header}

\item{\code{version}}{Las version information, as array}

\item{\code{data}}{\code{data.frame} object with tracks data, as data frame}

\item{\code{tracks}}{active binded field: tracks information}
}}

\keyword{datasets}
