% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{log_breaks}
\alias{log_breaks}
\title{Breaks For log10 Scale}
\format{An object of class \code{numeric} of length 21.}
\usage{
log_breaks
}
\description{
Breaks For log10 Scale
}
\keyword{datasets}
