% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{log_minor_breaks}
\alias{log_minor_breaks}
\title{Minor Breaks For log10 Scale}
\format{An object of class \code{numeric} of length 189.}
\usage{
log_minor_breaks
}
\description{
Minor Breaks For log10 Scale
}
\keyword{datasets}
