% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{rbind.las}
\alias{rbind.las}
\title{rbind() S3 Method - R6 Method caller}
\usage{
\method{rbind}{las}(..., deparse.level = 1)
}
\arguments{
\item{...}{N las object}

\item{deparse.level}{integer controlling the construction of labels in the case of non-matrix-like arguments (for the default method):
deparse.level = 0 constructs no labels; the default,
deparse.level = 1 or 2 constructs labels from the argument names, see the ‘Value’ section below.}
}
\description{
rbind() S3 Method - R6 Method caller
}
