% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petroreadr.R
\name{read_asciigrid}
\alias{read_asciigrid}
\title{Read grid properties exported in ascii format}
\usage{
read_asciigrid(filename, dims, index = TRUE, order = "IJK",
  along = "I")
}
\arguments{
\item{filename}{A character string giving the name or path of the file to be read.}

\item{dims}{An array of length = 3,telling the IJK dimension of the cube.
Has no default.}

\item{index}{A logic flag telling if IJK index should be returned or not.
If set to \code{FALSE} following parameters are ignored.
Default \code{TRUE}.}

\item{order}{A character string defining in which order Petrel exported the grid.
Petrel Export Options:
Cell Origin at (I=0, J=0, K)
Cell Origin at (I=0, J=max(J), K)
Cell Origin at (I=max(I), J=max(J), K)
Cell Origin at (I=max(I), J=0, K)
Should be one of: "IJK","IJmaxK","ImaxJmaxK" or "ImaxJK".
Default \code{"IJK"}.}

\item{along}{A character string defining in which order traverse the indexes.
Traverse first along I, then along J
Traverse first along J, then along I
Should be either: "I" or "J"
Default \code{"I"}.}
}
\value{
This function returns a dataframe with the grid properties and for
  each cell, and if \code{index = TRUE} the I, J and K indexes of the cells.
}
\description{
This function read 3D grid properties exported in ascii format and returns a
data frame with the resepctive properties value, and optionallythe I, J and K
indexes.
}
\details{
If the \code{filename} you provide does not existe, the function stops and
returns an error message.
}
\note{
Basically what this function do is to convert the ascci format into a
  easier to operate gslib-like format.

At the moment the function only read 1 porperty per file.
  Upgrade to read several properties in the same file is under development.
}
\examples{
\dontrun{
  read_asciigrid("filename.las", dims = c(12, 12, 25))
  read_asciigrid("filename.las", index = FALSE)
}

}
