% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petroreadr.R
\name{read_gslib}
\alias{read_gslib}
\title{Read .gslib grid format}
\usage{
read_gslib(filename, complete = TRUE, propnull = "-99.00")
}
\arguments{
\item{filename}{A character string giving the name or path of the file to be read.}

\item{complete}{A logic flag telling if complete cases only should be returned.
Default \code{TRUE}.}

\item{propnull}{A character string specifying the null property value.
Default \code{"-99.00"}.}
}
\value{
This function returns a dataframe with lthe logs data read from a .las file.
}
\description{
This function read 3D grid properties in .gslib format and returns a data
frame.
}
\details{
If the \code{filename} you provide does not existe, the function stops and
return an error message.
}
\examples{
\dontrun{
  read_gslib("filename.gslib")
}

}
