% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petroreadr.R
\name{read_las}
\alias{read_las}
\title{Read .las files}
\usage{
read_las(filename, lasnull = "-999.250000")
}
\arguments{
\item{filename}{A character string giving the name or path of the file to be read.}

\item{lasnull}{A character string indicating which value is sued as \code{NULL}}
}
\value{
This function returns a dataframe with lthe logs data read from a .las file.
}
\description{
This function read loga data from a .las file and return a data frame with
depth and all log values.
}
\details{
If the \code{filename} you provide does not existe, the function stops and
return an error message.
}
\note{
read_las has been calibrated on .las files exported from Petrel software
(Slb). Files .las generated by other software is still being tested.
}
\examples{
\dontrun{
read_las("filename.las")
read_las("pathname/filename.las", "-999")
}

}
