% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{read_las_}
\alias{read_las_}
\title{Read .las logs file - INTERNAL FUNCTION}
\usage{
read_las_(filename, verbose = FALSE)
}
\arguments{
\item{filename}{A character string giving the name or path of the file to be read.}

\item{verbose}{logical, define if the functions print the name of the file read. Default = FALSE}
}
\value{
an R6 object of class \code{las} or \code{mlas}
}
\description{
Read .las logs file - INTERNAL FUNCTION
}
