% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{scale.las}
\alias{scale.las}
\title{scale() S3 Method - R6 Method caller}
\usage{
\method{scale}{las}(x, center = FALSE, scale = FALSE)
}
\arguments{
\item{x}{las object}

\item{center}{either a logical value or numeric-alike vector of length equal to the number of columns of x,
here ‘numeric-alike’ means that as.numeric(.) will be applied successfully if is.numeric(.) is not true.}

\item{scale}{either a logical value or a numeric-alike vector of length equal to the number of columns of x.}
}
\description{
scale() S3 Method - R6 Method caller
}
