% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{summary.las}
\alias{summary.las}
\title{Summarizing `las` Objects}
\usage{
\method{summary}{las}(object, ...)
}
\arguments{
\item{object}{las object}

\item{...}{additional arguments}
}
\description{
Summary method for class `las`. The function call the R6 method defined within the class
}
\examples{
\dontrun{
summary(obj)
obj$summary()
}
}
