% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary_mod}
\alias{summary_mod}
\title{Object Summaries Modified}
\usage{
summary_mod(object, ..., digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}

\item{digits}{integer, used for number formatting with signif() (for summary.default)
or format() (for summary.data.frame). In summary.default, if not specified
(i.e., missing(.)), signif() will not be called anymore (since R >= 3.4.0,
where the default has been changed to only round in the print and format methods).}
}
\value{
The form of the value returned by summary depends on the class of its argument.
   See the documentation of the particular methods for details of what is produced by that method.

   The default method returns an object of class c("summaryDefault", "table") which has specialized
   format and print methods. The factor method returns an integer vector.

   The matrix and data frame methods return a matrix of class "table", obtained by applying summary
   to each column and collating the results.
}
\description{
summary is a generic function used to produce result summaries of the results
of various model fitting functions. The function invokes particular methods
which depend on the class of the first argument.
}
\note{
this function is a modification of the \code{base::summary} function to add:
   - p90
   - p10
   - st.dev
}
\examples{
\dontrun{
summary(data, digits = 2)
}

}
