% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_las_R6.R
\name{tail.las}
\alias{tail.las}
\title{tail() S3 Method - R6 Method caller}
\usage{
\method{tail}{las}(x, n = 6L, ...)
}
\arguments{
\item{x}{las object}

\item{n}{integer, number of rows}

\item{...}{additonal arguments}
}
\description{
tail() S3 Method - R6 Method caller
}
