% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petroreadr.R
\name{write_las0}
\alias{write_las0}
\title{Write well log dataframe to las file format}
\usage{
write_las0(data, name, verbose = TRUE, do = FALSE)
}
\arguments{
\item{data}{a dataframe. The curves names are taken from the column names of the dataframe.}

\item{name}{variable to be used as .las file name. Usually in the form \code{name = unique(.$id)},
where \code{id} was used as grouping variable. The extension .las is added automatically.}

\item{verbose}{logical, define whether the function should print out the names of the writen .csv files.
Dafault = TRUE.}

\item{do}{logical, define if the output is a data frame. Dafault = FALSE.
This is required if used with function \code{dplyr::do}, which requires to return a dataframe.
The obtained dataframe can be dumped with \code{rm(dataframe_name)}.}
}
\value{
write .las file.
    Additionally returns a dataframe if \code{do = TRUE} .
    As side effect, if \code{verbose = TRUE} prints out the names of the writen .las files
}
\description{
Write well log dataframe to las file format
}
\examples{
\dontrun{
write_las(df , name = paste(unique(.$id), "_version01"))

dump_df <- df \%>\%
 dplyr::group_by(id) \%>\%
 do(write_las(. , name = unique(.$id), do = TRUE))
rm(dump_df)
}

}
