% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petroreadr.R
\name{read_las}
\alias{read_las}
\title{Read .las files}
\usage{
read_las(filename, lasnull = "-999.250000")
}
\arguments{
\item{filename}{A character string giving the name or path of the file to be read.}

\item{lasnull}{A character string indicating which value is sued as \code{NULL}}
}
\value{
This function returns a dataframe with the logs data read from a .las file.
   It include a first column with the wellname.
}
\description{
This function read well logs data from a .las file and return a data frame with
   depth and all log values.
   If the \code{filename} you provide does not existe, the function stops and
   return an error message.
}
\note{
\code{read_las} has been tested to LAS 2.0 standard, check:
    <http://www.cwls.org/wp-content/uploads/2017/02/Las2_Update_Feb2017.pdf>
}
\examples{
\dontrun{
read_las("filename.las")
read_las("pathname/filename.las", "-999")
}

}
